Combine jit and logger utilities
