+++ qc_lab/utils.py
@@ -1,0 +1,89 @@
+"""Utility helpers for optional JIT compilation and in-memory logging."""
+
+# Combine functionality from jit.py and logger_utils.py
+
+import logging
+from io import StringIO
+from ._config import DISABLE_NUMBA
+
+logger = logging.getLogger(__name__)
+
+# ----- JIT utilities -----
+
+def qc_lab_custom_jit(func=None, **kwargs):
+    """Dummy jit decorator that does nothing if numba is not available."""
+    logger.info(
+        "Numba is disabled; using dummy jit decorator.",
+    )
+
+    def decorator(func):
+        return func
+
+    if func is None:
+        return decorator
+    else:
+        return decorator(func)
+
+
+def qc_lab_custom_njit(func=None, **kwargs):
+    """Dummy njit decorator that does nothing if numba is not available."""
+    logger.info(
+        "Numba is disabled; using dummy njit decorator."
+    )
+
+    def decorator(func):
+        return func
+
+    if func is None:
+        return decorator
+    else:
+        return decorator(func)
+
+
+if DISABLE_NUMBA:
+    njit = qc_lab_custom_njit
+    jit = qc_lab_custom_jit
+else:
+    try:
+        from numba import jit, njit
+    except ImportError:
+        jit = qc_lab_custom_jit
+        njit = qc_lab_custom_njit
+
+
+# ----- Logging utilities -----
+
+_log_stream = StringIO()
+
+class QCDataHandler(logging.Handler):
+    """Logging handler that stores logs in memory."""
+
+    def emit(self, record):
+        msg = self.format(record)
+        _log_stream.write(msg + "\n")
+
+
+def configure_memory_logger(level=logging.INFO):
+    """Configure root logger to store logs without printing."""
+    root_logger = logging.getLogger()
+    root_logger.setLevel(level)
+    # Remove existing handlers to avoid duplicate output
+    for handler in list(root_logger.handlers):
+        root_logger.removeHandler(handler)
+    handler = QCDataHandler()
+    formatter = logging.Formatter("%(levelname)s:%(name)s:%(message)s")
+    handler.setFormatter(formatter)
+    root_logger.addHandler(handler)
+    return handler
+
+
+def get_log_output() -> str:
+    """Return all collected log messages."""
+    return _log_stream.getvalue()
+
+__all__ = [
+    "jit",
+    "njit",
+    "configure_memory_logger",
+    "get_log_output",
+]
